+++ examples/snake/index.html
@@ -10,7 +10,7 @@
   </head>
   <body>
     <div id="root"></div>
-    <script type="module" src="/src/logic.ts"></script>
+    <script type="module" src="/src/logic/logic.ts"></script>
     <script type="module" src="/src/client.tsx"></script>
   </body>
 </html>
